put log related code to tree-item-log.js
