+++ front-end/components/business-process/tree/tree-view.js
@@ -6,39 +6,9 @@ import TreeItem from '@mui/lab/TreeItem';
 import {BPTextButton} from '../common/button';
 import {sampleEAIDomains} from '../../../utils/business-process/sample-data';
 import {BPColors, BPDimens, BPStandards} from '../../../utils/business-process/standards';
-import Image from 'next/image';
+import renderBusinessProcessInstances from './tree-item-log';
 
 const contextMenu = (e, source) => console.log( source + 'rightclick');
-
-const getColorBySeverity = (severity) => {
-  switch (severity) {
-    case 'success':
-      return BPColors.success;
-    case 'info':
-      return BPColors.info;
-    case 'warning':
-      return BPColors.warning;
-    case 'error':
-      return BPColors.error;
-  }
-};
-
-const BPActivitySeverityIcons = {
-  info: '/business-process/icons/severity-icons-info.svg',
-  warning: '/business-process/icons/severity-icons-warning.svg',
-  error: '/business-process/icons/severity-icons-error.svg',
-  success: '/business-process/icons/severity-icons-success.svg',
-};
-
-const BPActivitySeverityIcon = ({severity}) => {
-  return (
-    <Image
-      width={18}
-      height={18}
-      src={BPActivitySeverityIcons[severity]}
-    />
-  );
-};
 
 const findExpandable = (tree) => {
   const result = [];
@@ -150,37 +120,6 @@ const renderBusinessProcesses = (nodes) => (
   </TreeItem>
 );
 
-const renderBusinessProcessInstances = (log) =>(
-  <TreeItem
-    key={log.id}
-    nodeId={log.id}
-    icon={<BPActivitySeverityIcon severity={log.severity}/>}
-    label={log.sampleContent}
-    sx={{
-      marginTop: '1px',
-      borderRadius: BPDimens.treeRadius,
-      color: getColorBySeverity(log.severity),
-      backgroundColor: BPColors.transparent,
-      '&:hover': {
-        backgroundColor: BPColors.gray[100],
-      },
-      '& > .MuiTreeItem-content': {
-        minHeight: 34,
-        borderRadius: BPDimens.treeRadius,
-        padding: '0px 13px',
-        '&.Mui-focused, &.Mui-selected, &.Mui-focused.Mui-selected': {
-          backgroundColor: BPColors.gray[100],
-          '&:hover': {
-            backgroundColor: BPColors.transparent,
-          },
-        },
-        '&:hover': {
-          backgroundColor: BPColors.transparent,
-        },
-      },
-    }}
-  />
-);
 
 const data = sampleEAIDomains;
 const _expandable = findExpandable(data);
